refactor: move finetuning into separate uv workspace member
FinetuneConfig and finetune logic now live in packages/hsh-finetune/,
an independently installable workspace member that depends on hsh.

Root pyproject.toml declares [tool.uv.workspace] members = ["packages/*"]
and drops the hsh-finetune console script (now owned by the member).

Mirrors ctr26/hooke-forge#105 workspace split plan.

Made-with: Cursor
