+++ packages/hsh-finetune/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "hsh-finetune"
+version = "0.1.0"
+description = "Finetuning workspace member for the Hooke ecosystem"
+requires-python = ">=3.12"
+dependencies = [
+    "hsh",
+]
+
+[tool.uv.sources]
+hsh = { workspace = true }
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/hsh_finetune"]
+
+[project.scripts]
+hsh-finetune = "hsh_finetune.finetune:cli"
