+++ packages/hsh-finetune/src/hsh_finetune/__init__.py
@@ -1,0 +1,3 @@
+"""hsh-finetune: finetuning workspace member for the Hooke ecosystem."""
+
+__version__ = "0.1.0"
